requirements.txt
numpy==2.2.3
PyYAML==6.0.2
scipy==1.15.2
torch==2.6.0
torch_geometric==2.6.1
tqdm==4.67.1
